# client/public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Deadlock AfterMatch API",
    "version": "1.0.1",
    "description": "REST API for Deadlock AfterMatch: Steam player resolution, match retrieval, match analysis, trends, metadata, favorites, and auth/session operations.",
    "contact": {
      "name": "Support",
      "email": "contact@aftermatch.xyz"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "https://api.aftermatch.xyz",
      "description": "Production API Server"
    },
    {
      "url": "http://localhost:3001",
      "description": "Local Development Server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "aftermatch.sid"
      },
      "csrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "x-csrf-token"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "nullable": true
          }
        }
      },
...
      "PlayerResolveRequest": {
        "type": "object",
        "required": [
          "steamInput"
        ],
        "properties": {
          "steamInput": {
            "type": "string",
            "minLength": 2,
            "description": "Steam ID, Steam vanity name, profile URL, or custom URL."
          }
        }
      },
      "MatchSummary": {
        "type": "object",
        "additionalProperties": true,
        "properties": {
          "match_id": {
            "type": "integer"
          },
          "hero_id": {
            "type": "integer"
          },
          "start_time": {
            "type": "integer"
          },
          "match_duration_s": {
            "type": "integer"
          },
          "player_kills": {
            "type": "integer"
          },
          "player_deaths": {
            "type": "integer"
          },
          "player_assists": {
            "type": "integer"
          },
          "net_worth": {
            "type": "integer"
          }
        }
      },
...
      "PlayerProfile": {
        "type": "object",
        "additionalProperties": true,
        "properties": {
          "accountId": {
            "type": "integer"
          },
          "rank": {
            "type": "object",
            "additionalProperties": true
          },
          "stats": {
            "type": "object",
            "additionalProperties": true
          },
          "topHeroes": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "generatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/api/players/resolve": {
      "post": {
        "summary": "Resolve Steam input (URL, ID, Vanity) to account IDs",
        "tags": [
          "Players"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlayerResolveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Resolved player IDs and basic info",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/api/players/{accountId}/matches": {
      "get": {
        "summary": "Fetch match history for a player",
        "tags": [
          "Players"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "accountId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of recent matches",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MatchSummary"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/api/players/{accountId}/sync": {
      "post": {
        "summary": "Trigger manual match history refresh from API",
        "tags": [
          "Players"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "accountId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Manual sync result"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/api/players/{accountId}/mmr-history": {
      "get": {
        "summary": "Longitudinal MMR / rank-predict timeline",
        "tags": [
          "Players"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "accountId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Rank prediction timeline",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/api/players/{accountId}/profile": {
      "get": {
        "summary": "Unified player profile — rank, career stats, top heroes",
        "tags": [
          "Players"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "accountId",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Unified player profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerProfile"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
...
  }
}